+++ server/src/util.ts
@@ -4,4 +4,8 @@ function sleep(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-export { stringify, sleep };
+function range(size: number, startAt: number = 0): ReadonlyArray<number> {
+  return [...Array(size).keys()].map((i) => i + startAt);
+}
+
+export { stringify, sleep, range };